+++ deepinv/physics/__init__.py
@@ -19,7 +19,7 @@ from .noise import (
     LogPoissonNoise,
     GammaNoise,
 )
-from .mri import MRI, DynamicMRI, SequentialMRI, MultiCoilMRI, MRIMixin
+from .mri import MRI, DynamicMRI, SequentialMRI, MultiCoilMRI
 from .tomography import Tomography
 from .lidar import SinglePhotonLidar
 from .singlepixel import SinglePixelCamera
@@ -30,7 +30,6 @@ from .phase_retrieval import (
     StructuredRandomPhaseRetrieval,
 )
 from .radio import RadioInterferometry
-from .time import TimeMixin
 from .structured_random import StructuredRandom
 
 from . import generator
